Add pixel art project
Load pixel art in from a file of rgb values, then translate it to a series of images on the game screen.

Not particularly efficient, but I don't think there's a particularly efficient way to do this.

Does not contain a way to generate a file of actual art turned into rgb values; contains a gradient generator, essentially.